+++ UnityAds/UnityAdsBannerCustomEvent.h
@@ -1,0 +1,13 @@
+#import <UnityAds/UADSBanner.h>
+
+#if __has_include(<MoPub/MoPub.h>)
+    #import <MoPub/MoPub.h>
+#elif __has_include(<MoPubSDKFramework/MoPub.h>)
+    #import <MoPubSDKFramework/MoPub.h>
+#else
+    #import "MPBannerCustomEvent.h"
+#endif
+
+@interface UnityAdsBannerCustomEvent : MPBannerCustomEvent<UnityAdsBannerDelegate>
+
+@end